Switch CArray to a pointer.
This makes basically all use cases look cleaner.

The only cons are that:
* the notation array->count is confusing if you forget array is a ptr;
* you have to malloc(sizeof(CArrayStruct)) to allocate one yourself.

I consider these worthwhile since 99% of the time, treating a CArray as
an opaque object is the right thing to do. We can often drop the stress
of worrying about levels of indirection that come with pointer use.