Add array.{h,c} toward renaming things.
This is the first step toward naming things with less camel casing,
and hopefully with more clarity. The word "element" is being replaced by
"item" as it's shorter. Functions that may work with either pointers to
items or direct item values use the strings "ptr" and "val" for added
clarity.